course decryption logic in its own hook
